Add pseudo-types for common things like mass, position, velocity, force, etc.
This is to make the code more expressive and get around the fact that named types aren't a thing. So leveraging the type system to make it a *bit* more clear about what is going on here. Downside here is these using delcarations aren't that strong. I could pass a velocity in as a mass or vice versa. But this is a good start for making way for ODE solvers.